+++ packages/component/src/page/trading/sections/navbar/marketOverview.tsx
@@ -11,8 +11,6 @@ export const MarketOverview: FC<Props> = (props) => {
   const { symbol } = props;
   const data = useTickerStream(symbol);
 
-  console.log("data", data?.symbol, data?.["24h_close"]);
-
   return (
     <SymbolProvider symbol={symbol}>
       <SimpleMarketOverview