+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @orderly.network/hooks
 
+## 0.0.81
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.74
+
+## 0.0.80
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.73
+
+## 0.0.79
+
+### Patch Changes
+
+- update
+- update
+- Updated dependencies
+  - @orderly.network/net@1.0.72
+
 ## 0.0.78
 
 ### Patch Changes