+++ packages/net/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/net
 
+## 1.0.74
+
+### Patch Changes
+
+- update
+
+## 1.0.73
+
+### Patch Changes
+
+- update
+
+## 1.0.72
+
+### Patch Changes
+
+- update
+
 ## 1.0.71
 
 ### Patch Changes